# Unroll the loop of string size calculation