@ohler55 2021-12-18
Nice improvement. As always, thank you for making Oj faster.